+++ test/app/app.js
@@ -1,5 +1,5 @@
 import {mdr} from '../../src/index.js'
 
-const input = '```gh-gist btnguyen2k/d7577db0981cb157ae95b67b9f6dd733\n```'
+const input = '```video invalid://file.mp4\n```'
 const output = mdr(input)
 console.log(output)